Store backedges in a Set
Summary:
I was able to find one place where we depend on the potentially inconsistent ordering of the backedges in D15461537.
However, there could be more of them. By storing the backedges as a set rather than a list, we can be sure we're operating on them in a consistent order, even if that order is arbitrary.
This would have been more straightforward had there not been a lot of code written for operations on both edges and backedges, which worked fine when they were both lists, but had to be functorized in order to support both interfaces.

Reviewed By: sinancepel

Differential Revision: D15416611

fbshipit-source-id: 15ff76050b790c461ec67d291534198c5f1db8a8