Change location of implicit return to last line of original function definition
Summary:
Originally, the implicit return location was the location of the entire function body, which is confusing because return statements are usually at the end of functions. To fix this, I recursively found the last statement of the function body by traversing the function definition tree and added its location as the location of the implicit return node inserted in expand_implicit_returns.

Additionally, I changed two tests in returnTest.ml to match the new order after the implicit return location was changed to the last line of the function.

Reviewed By: shannonzhu

Differential Revision: D15521051

fbshipit-source-id: 7b9d6d0ecfe557057b189271adf8fb3817d317c0