Add as_expression conversion to accessPath
Summary:
Need to supply this conversion in order to finish migrating callResolution off of the old access; currently it relies on callers converting accessPaths into accesses (and then adding on layers to turn it into a toplevel expression, anyway).

Since the new AST structure mimics the structure of normalized expressions, we can likely get rid of most of the extra conversion and scaffolding soon.

Reviewed By: fahndrich

Differential Revision: D15578400

fbshipit-source-id: 4cce17f20189fde84761cde4e2c898eafe7d3bff