Supporting list variadics in principle
Summary:
# Background

The meat of the work in this diff takes place in TypeConstraints.  For some context I'll give a brief overview of the architecture of that module.  The interface presented to a user is that of adding constraint pairs, and then solving them out.
A **constraint pair** consists of a variable and a member of its domain, and that can either be a lower bound or an upper bound.
For unary type variables this would be a variable and an upper bound type or a lower bound type.
For these new list variadics this is a list variadic variable and a Type.ordered_types variant which is either a ConcreteList of types, a single other list variadic, or a new concept, AnyList.
(AnyList was necessary to introduce in order to preserve the standard that trying to assign a type that contains free type variables converts those variables to Anys and emits an error).
These pairs refine the **interval** we define for any type variable.  An interval defines the set of members of a domain we have learned that a variable must be a member of.  For unaries and list variadics, this takes the form of upper and lower bounds that are each members of the domain.
Each new constraint pair is like a half interval, and then we intersect that interval with the existing one, if one exists.

To get a solution, for intervals which do not contain other variables, it's as easy as picking the bottom of the interval, presuming that the bottom is still less than the top of the interval.
For `dependent` intervals, we handle their dependency on other variables by substituting the partial solution we generate from these `independent` intervals into their bounds, and hope that they become independent.  If two variables are bounded by each other in any form, we decline to try to solve the system of constraints.

If all goes well this gives us a solution object, which is a combination of three maps from each kind of variable to their solved members of their respective domain.  This solution can be used to substitute out all matching free variables from a type.

# This diff
As discussed in detail in https://fburl.com/variadic ListVariadics are the next step in implementing variadic type variable functionality.
These are the variadic type variables which range over lists of types.
This immediate stack builds up to having them used in tuples in particular, but these will end up in a variety of what I call "hosts" for the type variable, including `*args` parameters and eventually user defined parametrics.
However this diff is completely agnostic to those hosts, and mostly is concerned with defining the constraint solving approach for these.

# Deeper details
I decided to go a little verbose with the `BothBounds` etc. approach.  Either a record like unary intervals with optional fields or explicitly created Top and Bottom for Type.ordered_types would have worked, but I rejected both of them for different reasons.
Optional upper and optional lower ended up being fairly confusing with later uses of optional sometimes corresponding to failure and other times to a lack of a bound.
Explicit top and bottom values were overkill IMO since they would need to be considered in all uses of Type.ordered_types when they could be contained within TypeConstraints.

Reviewed By: sinancepel

Differential Revision: D15484952

fbshipit-source-id: a4fc91b91c903645841b3bbafa25212158f1e64e