fix parse error for complex dictionary comprehension
Summary:
Historically, we've had our hand-rolled parser that mostly works good enough, but is quite divergent from the Python grammar specification. I ran into a parse error for the following snippet:

```
{killswitch.feature or "": killswitch for killswitch in killswitches}
```

I used this as an opportunity to make Pyre's dictionary or set maker specification closer to `https://docs.python.org/3/reference/grammar.html`, which is why I went with the names I chose to go with. Notable differences from the specification:

1. I went for left recursion instead of maker = (entry, maker), as from my skimming of the internet, it's a form of recursion that requires bounded space as opposed to right recursion.

2. I gave some names to parts of the description that the Python parser generator consumes.

3. I explicitly kept our own janky logic for handling `**expr` for dictionaries instead of fixing it. The way python's grammar suggests we do it is quite incompatible with the existing parser, and it seemed like a hill not worth dying on.

Gotchas:

I ran into a reduce/reduce conflict with the comprehension: rule and test: rule. Still not quite sure why, but breaking the `ASYNC?` rule into two rules helped fix it.

We still probably have parse errors for dictionaries as we parse things as "set or dictionary element", and try to figure what it was out later. This isn't great, but as mentioned above, we can't do better before fixing how we handle stars.

Reviewed By: dkgi

Differential Revision: D15546126

fbshipit-source-id: 22564dfb16ba917271ece1b74be76deeb3735b38