handle aliasing in awaitable check
Summary:
We previously assumed aliasing did not exist in awaitableCheck.

The basic case occuring in the wild that I aimed to handle was

```
a = awaitable()
b = a
await b
```

My original approach was to essentially use linear types, i.e., if we see the `b = a` assignment, assume `a` is awaited, and require `b` to be awaited afterwards. This doesn't work well in even slightly sophisticated cases.

Counter-example 1:

```
a = awaitable()
b = a
await a  # b is not awaited in our view
```

Counter-example 2:
```
a = awaitable()
b = a
c = a
await c  # b is not awaited
```

Observing this, go for the standard alias analysis approach: Introduce one level of indirection and maintain two separate maps: one from names to memory allocations, and one from memory allocations to awaitable tags.

I've chosen to represent memory allocations as the locals, i.e. targets of assignments, and awaitables as maps from the locations where awaitables are created to the unawaited expressions.

Reviewed By: shannonzhu

Differential Revision: D15636914

fbshipit-source-id: c68358926ea0e7abb37bd0956e5d397d67a3c77f