Add helper function to retrieve location of arguments with parenthesis
Summary:
In Pysa we care about the location of the arguments (the whole block, including parenthesis), rather than the location of the callee or any one particular argument.

The old AST used to keep the location of the argument block, but no other locations throughout an access chain. The new AST keeps locations for every attribute down the access chain as well as every argument value, but for match-case simplicity doesn't keep the argument block location anymore. Add some logic to fetch this location from the locations we do know.

Reference: D14842273

Reviewed By: dkgi

Differential Revision: D15584645

fbshipit-source-id: 694e4ba027db989a51a95f06cb6694c121003cab