Directly supporting keyword only parameters
Summary:
In anticipation of having `*args` be a host for ListVariadics, I am removing the `name` and `default` fields from Variable and Keyword arguments.
In order to do that though, we need to eliminate the one instance where we depend on the name of a variable parameter, namely our handling of keyword only parameters.
For background python since PEP 3102 supports parameters that can only be bound through keyword-style calling.  The syntax for this is `def foo(normal_param: ann, *, keyword_only_param: ann)`. We previously handled this by parsing that as a nameless variable parameter, and then special casing it in the signature select.  This was problematic as this special casing belonged in other places (like override validation).
This opens the door to simplifying our interface for callables more generally in the next diff.

Reviewed By: sinancepel

Differential Revision: D15611159

fbshipit-source-id: 7f6963a7aa11ff015557f4d536aa8972a9059a98