Simplifying the types of variable and kwargs parameters
Summary:
As promised in D15611159, in anticipation of making `*args` a host for `ListVariadic`s this diff removes the `name` and `default` fields from the `Variable` and `Keywords` parameter variants.  `default` is more or less a no-brainer since adding defaults to `*args` or `**kwargs` isn't even supported in the python grammar.
`name` is a bit subtler, but at the end of the day both of them are like anonymous parameters in that there is no way to directly address them by name, so the public interface of a callable type shouldn't include them.

Reviewed By: sinancepel

Differential Revision: D15628456

fbshipit-source-id: c5d681d76ea2ff973510e3553c3886803e055137