Eliminate use of diff variables from TypeOrder.ml
Summary:
`TypeOrder.diff_variables` is more or less a deeply unsound version of solve_less_or_equal.  Rather than updating it to be compatible with `Listvariadic`s in callables, I finally took this opportunity to try to remove it.

This diff has the simple replacements of diff_variables within TypeOrder itself, which straightforwardly were replaceable with a helper function that uses covariant tuples to do the same kind of comparison that was previously done with diff_variables.

Reviewed By: dkgi

Differential Revision: D15695318

fbshipit-source-id: 2438d6fb5c66dd1e2207db5b760592b743d21366