Correctly propagate protocol_assumptions back through typeorder callbacks
Summary:
The purpose of the protocol_assumptions set is to avoid infinite loops when determining whether a type implements a protocol depends on a type implementing a protocol.
Currently this was only being propagated internally within TypeOrder, but loops can also exist through the callbacks which typeorder is initialized with.
By propagating the protocol assumptions back out to the implementations of attributes, constructor and is_protocol we can safely call typeorder related functions without the risk of infinite loops.
These calls actually already exist, and as far as I can tell the only thing saving us from loops is that these situations are very hard to construct, but since loops are an awful, un-ignorable experience, it's worth coding defensively here.

Reviewed By: dkgi

Differential Revision: D15695382

fbshipit-source-id: fee46abbcc64155b2f1d9c61a03b511ff9e7b6ac