Alter expression printing to reflect special methods
Summary:
We already did some reconstruction when printing expressions in the case of `__getitem__`. Is this just because `__getitem__` is used widely in types, and things become very unreadable when we keep it in method-expanded form when printing?

Because this reconstruction exists already, all this diff does is limit it to when the original expression really did call `base[item]` rather than `base.__getitem__(item)` explicitly.

RFC:
Should I be reconstructing other special methods when printing, like `__add__`, `__rshift__`, etc.? Or maybe it would be most readable to leave it as-is and leave it up to whoever is looking into the expression to see what the original form was, if it's relevant?

Reviewed By: sinancepel

Differential Revision: D15716266

fbshipit-source-id: 0fa79a6fe66f5d74b3024dd1708d61f4c3a4e412