Error once per define on bad attribute accesses
Summary:
D15703320 introduced a lot of errors, often duplicates if the same attribute was being accessed or assigned to more than once over the span of a single function.

This de-dups those errors, also following a TODO comment that meant to do this.

Note:
```
class Foo:
    a: int = 1

Foo.a
Foo.b  # pyre-fixme
Foo.b
```

Will not (and should not) throw errors, since the error join happens before ignores are applied, and we don't want a plug-a-leak-causes-more-leaks error ignore experience for users.

Some reservations:
I am a bit worried that filtering errors like this could cause some confusion as to why one attribute access errors but another doesn't, especially if there is a lot of logic separating the two bad attribute accesses. Is the potential confusion / perceived lack of coverage worth the smaller error/ignore count? We don't do this with other kinds of errors if you write the same bad thing multiple times...

Reviewed By: MaggieMoss

Differential Revision: D15716077

fbshipit-source-id: aaebe227a293bb14f21d7e5486434958955a29f5