Clean up error join and filtering logic
Summary:
We were joining errors before we filter for suppression. This means that sometimes, errors could get joined into an error that winds up getting suppressed, and could create subtle bugs. Really, to avoid duplicate checking in error joins for errors that should be filtered, we should be filtering/suppressing before we do joins.

If all the errors that would have been joined are suppressed, then no behavior change -- but if only some of them would have been suppressed, we want to be joining and showing the non-suppressed ones.

Reviewed By: sinancepel

Differential Revision: D15718748

fbshipit-source-id: c94936826306bdd5874a3c2cce1dda50ba4780df