Lift define-specific type checking logic into a separate function
Summary:
This is the first diff in a series that allow us to schedule type-checking individual defines, which is a pre-requisite for fine-grained dependency. With this change the fixpoint queue management logic can be disentangled from the main type checking logic.

At the end of the series we'd like to eventually expose the functionality of the lifted `check_define` to other modules somehow. But before that some further cleanup work needs to be done. Stay tuned.

Reviewed By: sinancepel

Differential Revision: D15751785

fbshipit-source-id: 9666afc2de9929fe9cc5fd42194d0e8b081910da