Fix a corner case in assignment target unpacking
Summary:
We reported unpacking errors on examples like this:

```
def foo(x: Tuple[int]) -> None:
    a, *b = x
```

```
pyre: Unable to unpack single value, 2 were expected
```

But the code works fine at runtime. The problem is that we failed to consider the possibility that `*b` could be an empty list. This diff tries to fix it by assigning `b` the type of `Tuple[]`.

Reviewed By: sinancepel

Differential Revision: D15755226

fbshipit-source-id: 2bcddf9e96da1a5c1b2c89730d4da49a64a4a2aa