refactor StatusUpdate to not hardcode socket writes
Summary: The server's connection state should never be accessed without grabbing the lock. The intention of status update is to notify all persistent clients. Thankfully, the `broadcast_response` function does the work of broadcasting a response for us, so refactor broadcast_response to be outside the loop, only initialize StatusUpdate once, and broadcast to all persistent clients.

Reviewed By: dkgi

Differential Revision: D15753023

fbshipit-source-id: d4a54ea0f92642cdcf27ea22b1be9f86b02a1c29