make all elements of `state.connections` immutable
Summary:
It is not a good idea to allow members of state.connections to be mutable, as we intend for `state.connections` to only be accessed within a mutex, and allowing the hash set of persistent clients to escape doesn't allow us to enforce this invariant via the type system.

In order to support making state.connections consist of immutable elements, introduce Socket.Map and move persistent_clients over.

Reviewed By: shannonzhu

Differential Revision: D15752849

fbshipit-source-id: ae4d0c7aa608f9e3321aeb510cfdd33328b069f3