Tuple unpack cleanup [2/2]
Summary:
As promised in D15755226, let's clean up the tuple assignment unpack logic.

This diff removes the tuple length check in the outer pattern match and merges two pattern branches into one.

Note that we also have a subtle semantic change here: previously we emits an error when trying to unpack `Top` or `Any`. I think the behavior for `Any` is just incorrect and for `Top` I made the judgement call of not erroring on it since the generated error would get filtered out most of the time anyway.

Reviewed By: dkgi

Differential Revision: D15769284

fbshipit-source-id: 932450565970fb89c9b2116798b6666c3da1af3b