Request.process no longer takes socket as an argument
Summary:
While looking at the server code, I noticed that we passed in the socket to the request processing code. Having a long lived reference to a socket, which might change in the meantime, is not a good idea.

The only reason we needed the socket when processing was to respond to stop requests (as the server never returns from there). Accomplish the same idea by moving the write a layer up to the caller, and only have the logic of
stopping the server in Request.process.

Reviewed By: MaggieMoss

Differential Revision: D15760743

fbshipit-source-id: b9d1afa6796ada12df421debe5214de18ae432ad