Support parsing and building all python_wheel related targets
Summary:
A python wheel target composes of a `python_library` and `remote_file` (and `prebuilt_python_library` but it's not our concern).
I first modified the `BuildTargetsCollector` to make a first pass through the `python_library` that has `platform_deps` field to collect correct versions of remote python files, then in the second pass we only choose to parse those `remote_file` that matches the versions.

Reviewed By: dkgi

Differential Revision: D15749450

fbshipit-source-id: c711616e3db734efeca93fcfcbc9bf665cf34431