Don't print mismatches with any in location error diff
Summary:
Making location tests more readable - before, we count Location.any as equivalent to other locations when testing equality, but when printing the total structure we *do* want them to still print out as anys. We just don't want the error diff to also highlight these as differences, because then it can take forever to figure out exactly which location is off.

Note: `diff` supports a flag called `--ignore-matching-lines`, but the added and deleted line both has to match the regex for this to work. Also, the entire block needs to be ignored for any one line to be ignored. Turns out postprocessing the diff output is probably the most efficient way to accomplish this.

Reviewed By: dkgi

Differential Revision: D15767321

fbshipit-source-id: 5c1b82d5ade226cb32185bb10b1547412ec2bec6