Ensure all locations are checked for Anys in generatorTest
Summary:
We were only checking that expression nodes were not generated with an any location. Statements also have locations to check, as do some field types of expression.t.

I suspect there could be more of these nested locations introduced in the future as we push for more complete location typing on parts of the AST that map to something in user code.

Reviewed By: grievejia

Differential Revision: D15764719

fbshipit-source-id: 4d85d70a4eb1372f904ef6cf3ba009d5a9dd122a