Add dumping for AST with locations
Summary:
Add an interface (`pyre_dump_locations()`) for revealing the AST with locations attached. This is likely helpful for us, but also requested by those working on libCST so we can ensure our locations line up with one another.

Leveraging the ppx deriving yojson generated function seemed by far the cleanest way to accomplish an alternative set of printing rules throughout the entire AST, and JSON-ifying the output makes it easier to read and fold since it's less readable with more information when locations are attached.

Reviewed By: dkgi

Differential Revision: D15782365

fbshipit-source-id: 954caef3e4d14a2f61b49c93157acbc45dd97e18