Fix missing -remote.py problem for buck project builder
Summary: Previously, the buck builder will miss all generated thrift python code ending with `-remote` since it does not end with .py or .pyi extension. I rewrite the copying logic to specially handle the case for thrift code generation by renaming that file to make it end with `-remote.py`, so that the behavior is now consistent with the current default buck builder.

Reviewed By: sinancepel

Differential Revision: D15783968

fbshipit-source-id: c696218bd420fc2819f1f52792e7405f4579a03b