Fix python targets sources map key-value roles
Summary:
In the source mapping object of python targets, keys should be the destinations and values should be the original sources.
Previously, we mistakenly inverted their roles. After the fix, I discovered that we are creating some bad symbolic links because we do not ignore buck auto-generated source map entries like `{ ... "/path/to/dest" : "/base/path:target", ...`. I fixed this by choosing to nothing when we detect the source of symbolic link is not an actual existing file.

Reviewed By: sinancepel

Differential Revision: D15793015

fbshipit-source-id: 3bdcbd40e85596f9a6ed4790fef3c9e89a9fe8d7