separate the generation of implicit submodules from module definition
Summary: While investigating issues with placeholder stubs, I noticed that we were a) generating implicit submodules before another module could potentially be introduced, and b) had inconsistent local modes based on the module that created the submodule. Fix these issues, and pull the code out to emphasize that the submodules should be completely independent of the actually present modules.

Reviewed By: grievejia

Differential Revision: D15797325

fbshipit-source-id: 85fd4300c9cd3b3956f1c6fbe8b94b509f1ba3f9