Having AnnotatedCallable.create's solve_less_or_equal correctly handle Any
Summary:
This diff was the result of some extensive performance debugging to understand why D15695406 resulted in significantly more time on certain files.

The answer was that we were problematically handling a common** pattern like this

```
T = TypeVar("T")
TSelf = TypeVar("TSelf", bound="Query")
class Query(Generic[T]):
    def verbose_interface(self: TSelf, x: int, y: str) -> TSelf: ...
class CatQuery(Query[Cat]): ...

def use() -> None:
  CatQuery().verbose_interface(1, "A").verbose_interface(1, "B").verbose_interface(2, "A")...
```
In order to determine the type of `CatQuery.verbose_interface` we simulate a call to the self parameter, solving and propagating the type variables therein.
This means we're "passing" a `CatQuery <: Query[Cat]` to something expecting a `Query[Any]`.  This should be possible, but because of our policy to treat `Any` as `Top` until error filtering, this is actually false.  This means the `solve_less_or_equal` fails, and we don't replace the type variable.  This means that for every step in these verbose interface chain, we're generating escaped type variables, and the corresponding errors.  These have to be detected, errored, and replaced, then those errors have to be joined out etc.  Eventually these are filtered out because we're currently gating these errors, but the perf cost has already been spent.

In the pathological files, this was being multiplied by being evaluated `n` times for a callable with `n` overloads (for a function like `asyncio.gather` `n=7`).
I address this independent weakness in a follow up diff.

I think this highlights the overall danger of our approach to handling `Any`, i.e. the error first, filter later approach.  This has major perf consequences, as well as correctness ones (see T45656370).

Reviewed By: dkgi

Differential Revision: D15775651

fbshipit-source-id: d036643d02b35444de6a96a4b8bedc75c0275b7d