Only resolve types of arguments once
Summary:
This was a perf issue exposed in the investigation that resulted in D15775651.

As described there, we were re-resolving the type of a function argument once per parameter match.
This meant that we were re-evaluating lots of expensive functions once per overload, and potentially multiple times within a single overload in the case of `*args` and `**kwargs`.

I also think this makes signature select much more readable.

Reviewed By: grievejia

Differential Revision: D15764334

fbshipit-source-id: eaf8835db7c0a60eebce79dee56e518a6de761d1