Allow for declaration of functions with `*args: Ts`
Summary:
Pt. II / II of implementing variadic generics for variadic functions.

This diff  implements the syntax for declaring functions with a `*args` typed with a `ListVariadic`

# Declaration Syntax

```
Ts = pyre_extensions.ListVariadic("Ts")
def tuplefy(*args: Ts) -> typing.Tuple[Ts]:
   return args
```

Because `Ts` is not a type, this is unambiguously referring to a function which takes an unlimited number of arguments of particular types, then returns a bounded tuple of those particular arguments.  Inside of `tuplefy`, `args` is known to have the type `Tuple[Ts]`, which is what allows this function to be completely typechecked.

Reviewed By: dkgi

Differential Revision: D15683658

fbshipit-source-id: db4707eacc38edad3d2ddd1035a0851277849ac5