Fix path issue
Summary:
When running `pyre-infer` on a project that used a local config (like most of the fbcode projects). `pyre-infer` would add the stubs in the wrong place, throwing errors when you tried to apply them.

This adds logic so the types are added relative to where the `pyre_configuration.local` file is - and types can be applied without a workaround.

Reviewed By: grievejia

Differential Revision: D15721935

fbshipit-source-id: 1fcdd0ebc91ffe0820c8aa232e57af41c2be656e