make instantiate successors parameters avoid throwing
Summary:
This throwing interacted poorly with the way that we handle `x <: Union[...]`.
When comparing `T(bound)` against `Union[T, List[T]]`, we break that apart into
`T(bound)<: T` or `T(bound) <: List[T]`.
The first half works correctly and is what will give us the solution we want.  The other half throws an untracked exception because type variables are (correctly) not in the type order.  This gets caught outside of the or and so ruins the whole comparison.
Also in general there is really no reason to throw here, since we already happily return an optional.

Reviewed By: grievejia

Differential Revision: D15830822

fbshipit-source-id: 9b503b0dbf15b55b8f51ffa3da2bd145a5761db5