Prevent accidental source code override in buck builder
Summary:
Sometimes we might have two dependencies having the same output file, where one of the output file is produced by creating symbolic link, while the other one is created by unzipping a file. When the source files are different, and the symbolic link is created first, the unzipped file will override the original content of the source code, following the symlink.
The behavior is harmful and this commit disables it by not unzipping a file is the destination already has the file.

Reviewed By: sinancepel

Differential Revision: D15833104

fbshipit-source-id: b860dd3ed3298525cfd41419e1cb3e71662e2e2d