Consider `dataclasses.field` when generating `__init__` method for dataclasses
Summary:
It turns out that `dataclasses.field()` can be used to override the behavior of `dataclass` for one field. This diff tries to add support for parameter `init` (setting to `False` means the generated `__init__` should not include parameter for this field) and `default` (act as the default value for the corresponding `__init__` parameter when set).

Relevant Python doc: https://docs.python.org/3/library/dataclasses.html#dataclasses.field

Reviewed By: sinancepel

Differential Revision: D15796378

fbshipit-source-id: 2793d0826d21a942b6f1a0751e42146222aee3f5