remove most primitive substitutions
Summary:
This map predates our correct handling of referring to generic classes without parameters being corrected to filled parameters with Any.
This is now done in `Resolution.check_invalid_type_parameters` and is done for all generic classes, not just these few that were manually encoded into this map.

The builtin types are left in (e.g. `list`, `dict`, `type`) because they cannot actually be subscripted in the runtime.  Therefore the messaging around correcting them needs to be more extensive (e.g when using `list` it needs to suggest that you both use `typing.List` and also give it a parameter).  I'll address them in a followup.

Reviewed By: grievejia

Differential Revision: D15846362

fbshipit-source-id: 4f5caa199b49af4a8e080661affd4121abc950cc