don't traverse class hierarchy when storing callgraph
Summary:
This is adapting to what we do in the taint analysis. We don't store the attribute that we find when making the call (walking up the class hierarchy), instead we store a function that does not exists. E.g. with

```
class A:
  def foo():

class B(A): pass

def calls_B_foo(b: B):
  b.foo()  # attribute lookup is `A.foo` but we store `B.foo` in the dependency graph
```

The rationale for this is that when propagating taint we have to analyze `B.foo` and every overriding method of `B.foo`.

Don't think this applies for dependencies so we might want to store both.

Reviewed By: sinancepel

Differential Revision: D15818883

fbshipit-source-id: ba419792902d7e5994083bfe371b9061cceccd26