Don't aggressively split unions in solve_less_or_equal
Summary: By putting the union match before the free variable match, that meant that when solving statements like `Union[Tuple[int, str], Tuple[bool, None]] <: T` we would break apart the union then end up re-joining it together in the interval intersecting.  Because we eagerly union tuples of the same length to the pairwise union of their members, which is *an* upper bound of the two types, but is greater than the union, we lose information in this there-and-back-again.

Reviewed By: grievejia

Differential Revision: D15870942

fbshipit-source-id: dbcbd08c5521328b12e9d60cf7fddfe86bd859ba