Support source mapping case where one source is mapped to multiple output directory
Summary:
There is case where source A can be outputed at both path A and path B, but the current implementation will only choose one to output. I reversed the key and value of the map, so that one output path can be only mapped to one source, but no two sources can be mapped to one output path.
In the future, we should keep track of all output path that is mapped to by several sources and log warnings.

Reviewed By: sinancepel

Differential Revision: D15871416

fbshipit-source-id: 53434d7e06745c9dcbeda8d3577cef7be60cc28c