Only piping standard error for thrift_library and swig_library with non-zero exit
Summary:
Previously we pipe all standard error while building `thrift_library` and `swig_library`. There are a lot of warning messages printed in thrift files while buck does not print them.
I think it might be helpful to make the printing behavior be consistent with buck, so now it will only print standard error if the exit code of build command is non-zero.
It is observed that the performance gain/loss is negligible but we reduced a lot of noise during building time.

Reviewed By: sinancepel

Differential Revision: D15889640

fbshipit-source-id: ed467f681ba4d626fbd55abc25ffc4cf16a01e9b