Expose API to enable per-function type check scheduling
Summary:
This diff adds an API `TypeCheck.run_on_defines` that, given a source and a list of functions it contains, would only type check those functions (as well as other functions that gets nested inside) instead of going through the entire source. This should conclude all required functional changes for T45511678

The added API is intended to be useful for fine-grained dependencies, where in incremental mode we can partially check dependencies of the changed files. Since we don't have the dependency computation part ready yet, currently the API is not used anywhere :(

Reviewed By: shannonzhu

Differential Revision: D15869146

fbshipit-source-id: 96ea81d66bdf8b835ef66501e28c61446c9c4b29