Back out "[pyre] quick fix to avoid infinite loop"
Summary:
The original change was running into an infinite loop when recursively invoking `forward_expression` on the base of the attribute. Avoiding this by keeping this around in the state when resolving the attribute in the first place.

This is a bit invasive but really have no idea where the other issue is coming from and don't have the time to go on a week long debugging binge.

On the plus side: this is strictly more performant.

Reviewed By: shannonzhu

Differential Revision: D15893966

fbshipit-source-id: e6df8edb42e8899397e70f4449b5f85285449037