Supporting map operator inside of function declarations & callable types
Summary:
As promised by D15837611, the next home for `MapOperator` is inside of functions/callables.
This diff makes them available everywhere normal `ListVariadics` can be placed, namely as annotations for `*args` and inside of `Callable` types.
This completes what I've been calling "Phase 1" of the variadic type variable project, which constituted a complete solution for handling variadic functions.  The two main targets, `builtins.map` and `asyncio.gather` now have demonstrated replacements**.

** The replacement for `asyncio.gather` is not quite productionizable yet since the actual signature is
```
_FutureT = Union[Future[_T], Generator[Any, None, _T], Awaitable[_T]]
def gather(coro_or_future1: _FutureT[_T1],
            *, loop: AbstractEventLoop = ..., return_exceptions: bool = ...) -> Future[Tuple[_T1]]: ...
# etc...
```
This is using a generic alias which we do not support at all.  Rather than block deployment of this feature on implementing generic aliases, we could replace that alias with a class, and add it as a parent to each of those three classes.

Reviewed By: dkgi

Differential Revision: D15867265

fbshipit-source-id: 912ea47ce41763828499862c79ecbcb532d4ed34