Defining the union upper bound of ordered types more correctly
Summary:
After some contemplation, I realized that rounding a `Union[Map[List, Ts]]` to `List[object]` didn't make any sense due to variance.  This means we have to round both `Union[Ts]` and `Union[Map[List, Ts]]` just to `object`.  Eventually we should be able to make `Union` take an `OrderedTypes.t`, but until then, this works.

With this method in hand, I can now do basic propagation of these variadics for tuples, so that a Tuple[Ts] <: Iterable[object]

Note that I'm deferring the use of this new function on normal concrete tuples, as it would constitute a small functional change from using a join to just using a union.  The next diff does just that.

Reviewed By: dkgi

Differential Revision: D15882699

fbshipit-source-id: e6f6e2916e3b82934962b9d51a4783724bd55f75