Add new query for dependent defines
Summary:
This helps us to collect better metrics for fine-grained dependency work.

Currently we redirect the query to `Analysis.Dependencies.transitive_of_list`, grab the modules, and only counts the toplevel defines. Further improvement is certainly possible: it might be more informative if we include the nested defines as well. But we need to start somewhere.

Reviewed By: sinancepel

Differential Revision: D15905516

fbshipit-source-id: 0ef1995f15fcb9c153231fd2525e8b01a31dabd7