thrift_library building optimization by output generated code without copying
Summary:
I did a simple rewrite for thrift building command from -o [OUT] to -out [OUT], so that the generated sources will be put directly in [OUT] without /gen-py or /gen-py3 suffix. Therefore, we do not need to copy anything, so the related copying and temporary directory setup/cleanup code are both removed.
After this change, we have no control over what files to put into the output directory, so there are chances that generated thrift files can follow the symlink to override the original sources. Therefore, I move thrift library building to the first in the building pipeline to allow symlink creation to override the generated code.

Reviewed By: sinancepel

Differential Revision: D15923044

fbshipit-source-id: 9af6fbb5732d3dd536e4f816f2de755d5dda3539