don't ignore base methods when potential overrides exist
Summary:
Previously, when we had the following inheritance chain, `B.method` would only list `C.method` as a potential override:

```
class A:
  def method(self): ...
class B(A):
  pass
class C(B):
  def method(self): ...
```

This is a problem if `A.method` is tainted, as we still want to consider the case where `B.method()` might be a source of taint.

Change how CallResolution works so that we also consider `A.method`.

Reviewed By: dkgi

Differential Revision: D15905974

fbshipit-source-id: ae6507f1e9bf4f00c014700706a2761ddc2e7fae